Fixed broken ship movement (TODO: test for ship movement) Fixed broken adding cargo components to an entity didn't add extra storage. Changed cargo capacity and storage to a long rather than an int.

+++ Pulsar4X/Pulsar4X.CrossPlatformUI/Views/CargoView/CargoTypeStoreView.xeto.cs
@@ -32,7 +32,7 @@ namespace Pulsar4X.CrossPlatformUI.Views.CargoView
             CargoGrid.Columns.Add(new GridColumn
             {
                 HeaderText = "Amount",
-                DataCell = new TextBoxCell { ID = "AmountCell", Binding = Binding.Property<CargoItemVM, int>(r => r.Amount).Convert(r => r.ToString()) }
+                DataCell = new TextBoxCell { ID = "AmountCell", Binding = Binding.Property<CargoItemVM, long>(r => r.Amount).Convert(r => r.ToString()) }
             });
             CargoGrid.Columns.Add(new GridColumn
             {
